+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.3.0
+* The functions `tag!, tagsave` and their respective macros now obtain their arguments (besides the first two) as keywords instead of positional arguments. The positional versions are deprecated (#93).
+* New keyword `force = false` for `tag!` and co. which replaces the existing `gitcommit` field.
+
 # 1.2.0
 * Improved behavior of `savename` with respect to nested containers. If a nested container is empty, it is not printed instead. For example, `T=100_p=()_x=2` now becomes `T=100_x=2`. (if `p` is not empty then it is expanded as usual)